[boa-rollup-sdk] Separate internal modules into NPM modules
