+++ package.json
@@ -82,6 +82,7 @@
         "prettier": "^2.5.1",
         "prettier-plugin-solidity": "^1.0.0-beta.13",
         "prompt": "^1.3.0",
+        "rollup-pm-sdk": "^1.0.0",
         "smart-buffer": "^4.1.0",
         "solhint": "^3.3.6",
         "solidity-coverage": "^0.8.2",
