+++ src/modules/index.ts
@@ -1,13 +1,5 @@
-export { Hash, hash, hashMulti, hashFull, hashPart } from "./blockchain/common/Hash";
-export { Block } from "./blockchain/data/Block";
-export { BlockHeader } from "./blockchain/data/BlockHeader";
-export { Transaction } from "./blockchain/data/Transaction";
-export { TransactionPool } from "./blockchain/node/TransactionPool";
-
 export { NetworkError } from "./network/ErrorTypes";
 export { Scheduler } from "./scheduler/Scheduler";
 export { WebService } from "./service/WebService";
-export { JSONValidator } from "./utils/JSONValidator";
-export { Utils, ArrayRange, iota } from "./utils/Utils";
 
 export { IPFSManager } from "./network/IPFSManager";
